requirements.txt
django-environ
Django==5.2.6
djangorestframework==3.16.1
drf-spectacular==0.28.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
tqdm==4.67.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
